// src/constants/enums.ts
export enum SecureStoreEnum {
  AUTH_TOKEN_VALUE = '__AuthTokenSpeedCode__',
  AUTH_USER_INFO = '__AuthUserInfo__',
  AUTH_REFRESH_TOKEN = '__AuthRefreshToken__',
  AUTH_REFRESH_TOKEN_STORED_AT = '__AuthRefreshTokenStoredAt__',
  AUTH_LOGIN_EMAIL = '__AuthLoginEmail__'
}
export enum CacheDataKeyEnum {
  USER_PROFILE = 'USER_PROFILE',
  WALK_THROUGH_IGNORE = 'WALK_THROUGH_IGNORE'
}

export enum WebviewRouteEnum {
  test = '/my/test',
  advancedA = '/my/lesson/advanced-a/',
  advancedB = '/my/lesson/advanced-b/',
  meeting = '/meeting',
  chunk = '/my/chunk',
  seminar = '/my/seminar',
  navigate1 = '/navigation/1',
  navigate2 = '/navigation/2',
  faq = '/my/faq',
  setting = '/my/setting',
  password = '/my/password/',
  referral = '/my/referral/',
  interview = '/interview/',
  personal = '/service/personal/',
  record = '/my/lesson/record/',
  present = '/my/present/',
  download = '/my/lesson/download/',
  campaign = '/my/campaign',
  patapuraMethod = '/article/patapura-method',
  forgettingCurve = '/article/forgetting-curve',
}